Fix Validator handling for getOwnPropertyDescriptor on sealed properties
The Validator handler traps getOwnPropertyDescriptor and fowards the call
when there is no underlying trap. But in the case where the target has a
sealed property, the validator cannot return a descriptor that says the
property is sealed. This seems like a vague rule in the proxy specification.

The spec invariant states that a validator should change the configurable
property of the descriptor to true.